requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
google-genai==0.2.2
supabase==2.9.1
apscheduler==3.10.4
pytz==2023.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
flask-cors==4.0.0
pydantic==2.5.0
python-dateutil==2.8.2
secure==0.3.0